+++ src/LuigisBoxBundle/Factory/AbstractFactory.php
@@ -35,7 +35,7 @@ abstract class AbstractFactory
         return new Request(
             $this->getHttpMethod(),
             new Uri($this->configProvider->getHost() . $this->getEndpoint()),
-            $this->configProvider->getRequestHeaders($this->getHttpMethod(), $this->getEndpoint(), $now),
+            $this->configProvider->getAuthorizationHeaders($this->getHttpMethod(), $this->getEndpoint(), $now),
             $this->serializer->serialize($bodyObject)
         );
     }